@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
+    "@angular/animations": "^18.2.14",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.2.0",
-    "@angular/compiler": "^17.2.0",
-    "@angular/core": "^17.2.0",
-    "@angular/forms": "^17.2.0",
+    "@angular/common": "^18.2.14",
+    "@angular/compiler": "^18.2.14",
+    "@angular/core": "^18.2.14",
+    "@angular/forms": "^18.2.14",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.2.0",
-    "@angular/platform-browser-dynamic": "^17.2.0",
-    "@angular/router": "^17.2.0",
+    "@angular/platform-browser": "^18.2.14",
+    "@angular/platform-browser-dynamic": "^18.2.14",
+    "@angular/router": "^18.2.14",
     "@popperjs/core": "^2.11.8",
     "apexcharts": "^3.54.1",
     "bootstrap": "^5.3.3",
@@ -36,7 +36,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.2.2",
     "@angular/cli": "^20.2.0",
-    "@angular/compiler-cli": "^17.2.0",
+    "@angular/compiler-cli": "^18.2.14",
     "@schematics/angular": "^20.2.0",
     "@types/jasmine": "~5.1.0",
     "jasmine-core": "~5.1.0",
