@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import {HttpClient, HttpParams} from "@angular/common/http";
+import { HttpClient, HttpParams } from "@angular/common/http";
 import {Client} from "../../models/client";
 import {PaginatedResponse} from "../../models/paginatedResponse";
 
